Move all enums to enums module
